Variables
Variables allow you to assign and store value for a given piece of information. They are implicitly declared and strong-typed. Here some of the basic strong-typed variables you can store in your program. 

- Numeric
- String
- List 
- Tuple
- Dictionary

In the following pieces of code, Python will execute each line top to bottom, and assign the value on the left side of equal sign (assignment operator) to the variable on the right side of the operator.